backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
vaderSentiment 
openai
python-dotenv 